% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludePoints}
\alias{excludePoints}
\title{Filter points based on attribute values in a list}
\usage{
excludePoints(pts, excludelist)
}
\arguments{
\item{pts}{\code{sf} object containing points.}

\item{excludelist}{text file containing attribute name and values}
}
\value{
A \code{sf} object containing the filtered points.
}
\description{
Returns only the points that are not listed
}
