% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_points}
\alias{filter_points}
\title{Filter points that are parts of trajectories}
\usage{
filter_points(traj, pts, join_by)
}
\arguments{
\item{traj}{\code{sf} object containing the trajectories.}

\item{pts}{\code{sf} object containing points.}

\item{join_by}{shared attribute by which to filter.}
}
\value{
A \code{sf} object containing the filtered points.
}
\description{
Returns only the points that share a specified attribute with trajectories.
}
