% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterPoints}
\alias{filterPoints}
\title{Filter points that are parts of trajectories}
\usage{
filterPoints(traj, pts, joinBy)
}
\arguments{
\item{traj}{\code{sf} object containing the trajectories.}

\item{pts}{\code{sf} object containing points.}

\item{joinBy}{shared attribute by which to filter.}
}
\value{
A \code{sf} object containing the filtered points.
}
\description{
Returns only the points that share a specified attribute with trajectories.
}
