% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoints.R
\name{get_checkpoints}
\alias{get_checkpoints}
\title{Create checkpoint geometries}
\usage{
get_checkpoints(pts, x, y, size)
}
\arguments{
\item{pts}{\code{sf} object containing points.
include.}

\item{x}{\code{vector} of x coordinates of the checkpoint center.}

\item{y}{\code{vector} of y coordinates of the checkpoint center.}

\item{size}{Radius of the checkpoints.}
}
\value{
\code{sf} object containing the input points along with their closest
checkpoint number.
}
\description{
Create checkpoint geometries
}
