% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_overall_ldm}
\alias{get_overall_ldm}
\title{Calculate the linear directional mean of all trajectories combined}
\usage{
get_overall_ldm(traj)
}
\arguments{
\item{traj}{\code{sf} object containing trajectories.}
}
\value{
the linear directional mean in degrees over all given trajectories.
}
\description{
Calculate the linear directional mean of all trajectories combined
}
