% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{getPostgreSQLtbl}
\alias{getPostgreSQLtbl}
\title{Fetches data from a PostgreSQL table returns it as a data.frame}
\usage{
getPostgreSQLtbl(credentialFile, tblName, outFileName, bbx = NULL,
  ask = TRUE)
}
\arguments{
\item{credentialFile}{A csv file containing the database credentials}

\item{tblName}{Name of the table to be downloaded}

\item{outFileName}{Name of the output file}

\item{ask}{if \code{TRUE}, ask user for confirmation after showing the table
size}

\item{bbox}{Optional bounding box. A numeric \code{vector} of length 4 with
\code{xmin}, \code{ymin}, \code{xmax} and \code{ymax}.}
}
\description{
Fetches data from a PostgreSQL table returns it as a data.frame
}
