% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{make_arrows}
\alias{make_arrows}
\title{Make each trajectory a straight line of its cummulated length in its general
direction}
\usage{
make_arrows(traj)
}
\arguments{
\item{traj}{\code{sf} object containing trajectories.}
}
\value{
\code{sf} object containing straight lines.
}
\description{
Make each trajectory a straight line of its cummulated length in its general
direction
}
