% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{make_trajectories}
\alias{make_trajectories}
\title{Creates a set of trajectories from a sfc of points}
\usage{
make_trajectories(pts, foi, order_by, n)
}
\arguments{
\item{pts}{\code{sf} object containing points}

\item{foi}{Feature of interest by which the different trajectories are
differentiated}

\item{order_by}{Field by which points are ordered}

\item{n}{minimum number of points per trajectory}
}
\value{
A \code{sf} object containing the trajectories
}
\description{
Creates a set of trajectories from a sfc of points
}
