% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{makeMovementIndices}
\alias{makeMovementIndices}
\title{Calculates a number of movement indices for each trajectory}
\usage{
makeMovementIndices(traj)
}
\arguments{
\item{traj}{\code{sf} object containing the trajectories}
}
\value{
A \code{sf} object containing the trajectories with additional fields
}
\description{
Calculates average distance travelled per point and great circle distance
covered between start and end point for each trajectory in metres
}
