% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{makeTrajectories}
\alias{makeTrajectories}
\title{Creates a set of trajectories from a sfc of points}
\usage{
makeTrajectories(pts, foi, orderBy, n)
}
\arguments{
\item{foi}{Feature of interest by which the different trajectories are
differentiated}

\item{orderBy}{Field by which points are ordered}

\item{n}{minimum number of points per trajectory}

\item{\code{sf}}{object containing points}
}
\value{
A \code{sf} object containing the trajectories
}
\description{
Creates a set of trajectories from a sfc of points
}
