% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arrow_map.R
\name{plot_arrow_map}
\alias{plot_arrow_map}
\title{Plot trajectory data as a shiny leaflet arrow map}
\usage{
plot_arrow_map(traj)
}
\arguments{
\item{traj}{\code{sf} object containing trajectory data to display.}
}
\description{
Plot trajectory data as a shiny leaflet arrow map
}
