% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_ldm}
\alias{plot_ldm}
\title{Generates a polar plot showing the linear directional mean of all
trajectories.}
\usage{
plot_ldm(traj, y, size)
}
\arguments{
\item{traj}{\code{sfc} containing trajectories and a ldm attribute.}

\item{y}{The parameter displayed on the y axis.}

\item{size}{The attribute defining the point size in the plot.}
}
\description{
Generates a polar plot showing the linear directional mean of all
trajectories.
}
