% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_static}
\alias{plot_static}
\title{Generate a static map of the given set of trajectories}
\usage{
plot_static(traj, buffer = 0)
}
\arguments{
\item{traj}{\code{sfc} containing trajectories and a ldm attribute.}

\item{buffer}{Buffer of area to include around the trajectories.}
}
\description{
Generate a static map of the given set of trajectories
}
