% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotMap}
\alias{plotMap}
\title{Plots trajectory and point data as a shiny leaflet map}
\usage{
plotMap(traj, pts)
}
\arguments{
\item{traj}{\code{sf} object containing trajectory data to display}

\item{pts}{\code{sf} object containing point data to display}
}
\description{
Plots trajectory and point data as a shiny leaflet map
}
