% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotStatic}
\alias{plotStatic}
\title{Plots trajectory and point data as a static osmplotr map}
\usage{
plotStatic(traj, pts, reuseMap = TRUE)
}
\arguments{
\item{traj}{\code{sf} object containing trajectory data to display}

\item{pts}{\code{sf} object containing point data to display}

\item{reuseMap}{If \code{TRUE}, will try to reuse a background map stored in
\code{.tmp.map.png} from earlier function calls.}
}
\description{
Plots trajectory and point data as a static osmplotr map
}
