% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{postgres2gpkg}
\alias{postgres2gpkg}
\title{Fetches data from a PostGIS table and writes it to a local Geopackage file}
\usage{
postgres2gpkg(credentialFile, tblName, outFileName, ask = TRUE)
}
\arguments{
\item{credentialFile}{A csv file containing the database credentials}

\item{tblName}{Name of the table to be downloaded}

\item{outFileName}{Name of the output file}

\item{ask}{if \code{TRUE}, ask user for confirmation after showing the table
size}
}
\description{
Rows without geometries will be omitted.
}
