% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{postgres2sqlite}
\alias{postgres2sqlite}
\title{Fetches data from a PostgreSQL table and writes it to a local SQLite file}
\usage{
postgres2sqlite(credentialFile, tblName, outFileName, bbx = NULL,
  ask = TRUE)
}
\arguments{
\item{credentialFile}{A csv file containing the database credentials}

\item{tblName}{Name of the table to be downloaded}

\item{outFileName}{Name of the output file}

\item{bbx}{Optional bounding box. A numeric \code{vector} of length 4 with
\code{xmin}, \code{ymin}, \code{xmax} and \code{ymax}.}

\item{ask}{if \code{TRUE}, ask user for confirmation after showing the table
size}
}
\description{
Fetches data from a PostgreSQL table and writes it to a local SQLite file
}
