% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_credentials}
\alias{read_credentials}
\title{Read Database Credentials}
\usage{
read_credentials(f_name)
}
\arguments{
\item{f_name}{Filename of the csv file containing the credentials}
}
\value{
A \code{data.frame} containing the credentials
}
\description{
Reads a csv containing database information and returns them as a data.frame.
}
