% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{read_gpkg}
\alias{read_gpkg}
\title{Read GeoPackage file and return a spatial data.frame}
\usage{
read_gpkg(f_name, bbox = NULL)
}
\arguments{
\item{f_name}{Name of the GPKG file.}

\item{bbox}{a numeric \code{vector} of length 4 with \code{xmin},
\code{ymin}, \code{xmax} and \code{ymax}.}
}
\value{
A \code{data.frame} containing the specified table contents.
}
\description{
Read GeoPackage file and return a spatial data.frame
}
