% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{read_sqlite}
\alias{read_sqlite}
\title{Read SQLite file and return a data.frame}
\usage{
read_sqlite(f_name, tbl_name, sf = TRUE, bbox = NULL)
}
\arguments{
\item{f_name}{Name of the SQLite file.}

\item{tbl_name}{Name of the table to be fetched.}

\item{sf}{if \code{TRUE}, returns a \code{sf} object containing only data
with valid coordinates in fields lat and lon.}

\item{bbox}{a numeric \code{vector} of length 4 with \code{xmin},
\code{ymin}, \code{xmax} and \code{ymax}.}
}
\value{
A \code{data.frame} containing the specified table contents.
}
\description{
Read SQLite file and return a data.frame
}
