% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCredentials}
\alias{readCredentials}
\title{Read Database Credentials}
\usage{
readCredentials(fName)
}
\arguments{
\item{fName}{Filename of the csv file containing the credentials}
}
\value{
A \code{data.frame} containing the credentials
}
\description{
Reads a csv containing database information and returns them as a data.frame.
}
