% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{readSQLite}
\alias{readSQLite}
\title{Read SQLite file and return a data.frame}
\usage{
readSQLite(fName, tblName, sf = TRUE, bbox = NULL)
}
\arguments{
\item{fName}{Name of the SQLite file}

\item{tblName}{Name of the table to be fetched}

\item{sf}{if \code{TRUE}, returns a \code{sf} object containing only data
with valid coordinates in fields lat and lon}

\item{bbox}{a numeric \code{vector} of length 4 with \code{xmin},
\code{ymin}, \code{xmax} and \code{ymax}.}
}
\value{
A \code{data.frame} containing the entire table
}
\description{
Read SQLite file and return a data.frame
}
