% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{write_gpkg}
\alias{write_gpkg}
\title{Write spatial data object to GeoPackage file}
\usage{
write_gpkg(dat, out_file_name)
}
\arguments{
\item{dat}{\code{sf} object containing spatial data.}

\item{out_file_name}{File name of the saved object.}
}
\description{
Write spatial data object to GeoPackage file
}
