% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_ldm}
\alias{plot_ldm}
\title{Generates a polar plot showing the linear directional mean and trajectory
length.}
\usage{
plot_ldm(traj)
}
\arguments{
\item{traj}{\code{sfc} containing trajectories.}
}
\description{
Generates a polar plot showing the linear directional mean and trajectory
length.
}
