% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_ldm}
\alias{plot_ldm}
\title{Generates a polar plot showing the linear directional mean and trajectory
length.}
\usage{
plot_ldm(traj, x, y, size)
}
\arguments{
\item{traj}{\code{sfc} containing trajectories.}

\item{x}{The parameter displayed on the x axis.}

\item{y}{The parameter displayed on the y axis.}

\item{size}{The attribute represented by dot size or a fixed value.}
}
\description{
Generates a polar plot showing the linear directional mean and trajectory
length.
}
