% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{readSQLite}
\alias{readSQLite}
\title{Read SQLite file and return a data.frame}
\usage{
readSQLite(fName, tblName)
}
\arguments{
\item{fName}{Name of the SQLite file}

\item{tblName}{Name of the table to be fetched}
}
\value{
A \code{data.frame} containing the entire table
}
\description{
Read SQLite file and return a data.frame
}
